chore(releases): moving bundles/hooks to releases/hooks
