+++ packages/sanity/src/core/bundles/index.ts
@@ -1,3 +1,2 @@
-export * from './hooks'
 export * from './util/const'
 export * from './util/util'
